Removed generators that are moving to plugin gems and added zeitwerk (#71)
