+++ CHANGELOG.md
@@ -13,18 +13,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adopted Semantic Versioning
 - Added pa11y automated accessibility testing
 - Added dark mode support to our GitHub pages theme
-- Added alex inclusive language testing
 - Decided on Cabin as our Software-as-a-Service (SaaS) Web Analytics Solution
-- Added a Decision Record issue template
 - Added a panel for README badges and an HDI Way of Working badge
-- Added a Pull Request template and communications guidelines
-- Added a document command to update the list of used and available linters
-- Added Swiftlint build phase to Xcode projects
 - Added a plugin system for new functionality
 
 ### Change
 
 - Bumped GitHub Pages gem dependencies
+
+### Removed
+
+- MADR Decision Records moved to a separate plugin
+- MegaLinter moved to a separate plugin
+- Contributor Covenant Code of Conduct to a separate plugin
 
 ### Security
