--- docs/code-of-conduct.md
@@ -1,30 +1,0 @@
----
-layout: page
----
-
-# Code of Conduct
-
-We have adopted the [Contributor Covenant](https://www.contributor-covenant.org/version/2/1/code_of_conduct/) Code of Conduct v2.1
-
-Using the Contributor Covenant Code of Conduct benefits any community or project by setting clear expectations for acceptable behaviour and creating a safer and more inclusive environment for everyone involved. The Code of Conduct establishes guidelines for respectful communication and conduct and consequences for inappropriate behaviour. By adhering to these guidelines, community members can build trust and mutual respect, which fosters a more collaborative and productive environment.
-
-<!--alex disable sexual-->
-The Contributor Covenant is specifically designed to promote diversity and inclusion within communities and projects. It includes provisions to prevent discrimination and harassment based on factors such as race, gender, sexual orientation, and ability. By adopting and enforcing the Code of Conduct, communities can create a more welcoming and supportive environment for underrepresented groups.
-
-In addition to promoting a positive and inclusive environment, the Contributor Covenant can help prevent conflicts and misunderstandings by providing a clear set of expectations for behaviour, reducing the likelihood of interpersonal conflicts and creating a more harmonious community. Projects and communities can establish a strong foundation for healthy collaboration and growth using the Contributor Covenant Code of Conduct.
-
-{: .important }
-Adopting the Contributor Covenant Code of Conduct involves more than only adding a `CODE_OF_CONDUCT.md` file to the root of your project. Please thoroughly read the [enforcement guidelines](https://www.contributor-covenant.org/version/2/1/code_of_conduct/#enforcement-guidelines) section of the code and discuss the implications of adopting the code of conduct with your team.
-
-## Usage
-
-### Adding the Code of Conduct to your project
-
-To add a [Contributor Covenant](https://www.contributor-covenant.org/version/2/1/code_of_conduct/) Code of Conduct v2.1 to your project, use the following at the command line:
-
-    way_of_working init code_of_conduct --contact-method [CONTACT METHOD]
-
-You will need to enter a contact method, usually an email address.
-
-{: .note }
-We recommend that you use the HDI code of conduct email address as the contact method.
