--- docs/inclusive-language.md
@@ -1,28 +1,0 @@
----
-layout: page
----
-
-# Inclusive Language
-
-We have adopted [alex](https://alexjs.com) for automated testing of inclusive language.
-
-Using insensitive and inconsiderate language can cause harm to others, create barriers to communication, and damage relationships. It can make people feel excluded, disrespected, and devalued and may perpetuate negative stereotypes and biases. When we use insensitive or offensive language, we undermine our ability to connect with others and build trust.
-
-Therefore, it is essential to be mindful of the impact of our words and to choose language that is respectful, inclusive, and considerate of others. Doing so can create a more welcoming and inclusive environment for everyone, fostering greater understanding, empathy, and cooperation.
-
-To quote from the alex README:
-
-> No automated tool can replace studying inclusive communication and listening to the lived experiences of others. **An error from alex can be an invitation to learn more**.
-
-{: .important }
-Please take the opportunity to read some of the [further reading](https://github.com/get-alex/alex#further-reading) they reference and click on the links against individual warnings when they appear.
-
-## Usage
-
-To add alex to your project, run the following at the command line:
-
-    way_of_working init inclusive_language
-
-to run alex in your project, run:
-
-    way_of_working exec inclusive_language
