+++ lib/way_of_working/sub_commands/exec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require_relative 'base'
+
+module WayOfWorking
+  module SubCommands
+    # This class defines the exec (i.e. run) parent command
+    class Exec < Base
+    end
+  end
+end
