+++ lib/way_of_working/sub_commands/new.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require_relative 'base'
+
+module WayOfWorking
+  module SubCommands
+    # This class defines the "way_of_working new ..." sub command
+    class New < Base
+    end
+  end
+end
