--- lib/way_of_working/templates/.alexrc
@@ -1,3 +1,0 @@
-{
-  "profanitySureness": 1
-}
