--- lib/way_of_working/templates/.github/ISSUE_TEMPLATE/decision-record.md
@@ -1,26 +1,0 @@
----
-name: Decision Record
-about: Create a ticket to discuss or propose a new decision record
-title: ''
-labels: decision record
-assignees: ''
-
----
-## Context and Problem Statement
-
-{Describe the context and problem statement, e.g., in free form using two to three sentences or in the form of an illustrative story.
- You may want to articulate the problem in form of a question and add links to collaboration boards or issue management systems.}
-
-<!-- This is an optional element. Feel free to remove. -->
-## Decision Drivers
-
-* {decision driver 1, e.g., a force, facing concern, …}
-* {decision driver 2, e.g., a force, facing concern, …}
-* … <!-- numbers of drivers can vary -->
-
-## Considered Options
-
-* {title of option 1}
-* {title of option 2}
-* {title of option 3}
-* … <!-- numbers of options can vary -->
