--- lib/way_of_working/templates/.github/linters/.markdown-link-check.json
@@ -1,13 +1,0 @@
-{
-    "ignorePatterns": [
-        {
-            "pattern": "^0005-example.md"
-        },
-        {
-            "pattern": "^adr-template.md"
-        }
-    ],
-    "retryOn429": true,
-    "retryCount": 5,
-    "aliveStatusCodes": [ 200, 203 ]
-}
