--- lib/way_of_working/templates/.github/workflows/inclusive-language.yml
@@ -1,20 +1,0 @@
----
-name: "Inclusive Language"
-on: push
-
-permissions: read-all
-
-jobs:
-  test_accessibility:
-    name: Inclusive Language Tests
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v3
-
-      - uses: actions/setup-node@v2
-        with:
-          node-version: 18
-
-      - name: Run alex
-        run: npx alex --why
