--- lib/way_of_working/templates/.mega-linter.yml
@@ -1,129 +1,0 @@
----
-# Configuration file for MegaLinter
-# See all available variables at https://megalinter.io/configuration/ and in linters documentation
-
-APPLY_FIXES: none  # all, none, or list of linter keys
-ENABLE_LINTERS:  # If you use ENABLE_LINTERS variable, all other linters will be disabled by default
-  ### Languages
-  - BASH_EXEC  # in Super-Linter
-  - BASH_SHELLCHECK  # in Super-Linter
-  - BASH_SHFMT  # in Super-Linter
-  - C_CPPLINT
-  # - CLOJURE_CLJ_KONDO  # in Super-Linter
-  - COFFEE_COFFEELINT  # in Super-Linter
-  - CPP_CPPLINT  # in Super-Linter
-  - CSHARP_DOTNET_FORMAT  # in Super-Linter
-  - DART_DARTANALYZER  # in Super-Linter
-  - GO_GOLANGCI_LINT  # in Super-Linter
-  - GO_REVIVE
-  # - GROOVY_NPM_GROOVY_LINT  # in Super-Linter
-  - JAVA_CHECKSTYLE  # in Super-Linter
-  - JAVA_PMD
-  - JAVASCRIPT_ES  # in Super-Linter
-  # - JAVASCRIPT_PRETTIER
-  - JAVASCRIPT_STANDARD  # in Super-Linter
-  - JSX_ESLINT
-  - KOTLIN_KTLINT  # in Super-Linter
-  # - LUA_LUACHECK  # in Super-Linter
-  - MAKEFILE_CHECKMAKE
-  # - PERL_PERLCRITIC  # in Super-Linter
-  # - PHP_PHPCS  # in Super-Linter
-  # - PHP_PHPLINT  # in Super-Linter
-  # - PHP_PHPSTAN  # in Super-Linter
-  # - PHP_PSALM  # in Super-Linter
-  # - POWERSHELL_POWERSHELL
-  # - POWERSHELL_POWERSHELL_FORMATTER
-  - PYTHON_BANDIT
-  - PYTHON_BLACK  # in Super-Linter
-  - PYTHON_FLAKE8  # in Super-Linter
-  - PYTHON_ISORT  # in Super-Linter
-  - PYTHON_MYPY
-  - PYTHON_PYLINT  # in Super-Linter
-  - PYTHON_PYRIGHT
-  - R_LINTR  # in Super-Linter
-  # - RAKU_RAKU  # in Super-Linter
-  # - RUBY_RUBOCOP  # in Super-Linter
-  - RUST_CLIPPY  # in Super-Linter
-  # - SALESFORCE_SFDX_SCANNER_APEX
-  # - SALESFORCE_SFDX_SCANNER_AURA
-  # - SALESFORCE_SFDX_SCANNER_LWC
-  - SCALA_SCALAFIX
-  - SQL_SQL_LINT  # in Super-Linter
-  - SQL_SQLFLUFF  # in Super-Linter
-  - SQL_TSQLLINT
-  - SWIFT_SWIFTLINT
-  - TSX_ESLINT
-  - TYPESCRIPT_ES  # in Super-Linter
-  # - TYPESCRIPT_PRETTIER
-  - TYPESCRIPT_STANDARD  # in Super-Linter
-  # - VBDOTNET_DOTNET_FORMAT
-
-  ### Formats
-  # - CSS_SCSS_LINT  # Deprecated in favour of stylelint
-  - CSS_STYLELINT  # in Super-Linter
-  - ENV_DOTENV_LINTER  # in Super-Linter
-  # - GRAPHQL_GRAPHQL_SCHEMA_LINTER
-  # - HTML_DJLINT # Refuses to see config file
-  - HTML_HTMLHINT  # in Super-Linter
-  - JSON_ESLINT_PLUGIN_JSONC  # in Super-Linter
-  - JSON_JSONLINT
-  - JSON_NPM_PACKAGE_JSON_LINT
-  # - JSON_PRETTIER
-  - JSON_V8R
-  # - LATEX_CHKTEX  # in Super-Linter
-  - MARKDOWN_MARKDOWNLINT  # in Super-Linter
-  - MARKDOWN_MARKDOWN_LINK_CHECK
-  - MARKDOWN_MARKDOWN_TABLE_FORMATTER
-  - MARKDOWN_REMARK_LINT
-  # - PROTOBUF_PROTOLINT  # in Super-Linter
-  # - RST_RSTCHECK
-  # - RST_RSTFMT
-  # - RST_RST_LINT
-  - XML_XMLLINT  # in Super-Linter
-  # - YAML_PRETTIER
-  - YAML_V8R
-  - YAML_YAMLLINT  # in Super-Linter
-
-  ### Tooling
-  - ACTION_ACTIONLINT  # in Super-Linter
-  - ANSIBLE_ANSIBLE_LINT  # in Super-Linter
-  - ARM_ARM_TTK  # in Super-Linter
-  # - BICEP_BICEP_LINTER
-  - CLOUDFORMATION_CFN_LINT  # in Super-Linter
-  - DOCKERFILE_HADOLINT  # in Super-Linter
-  - EDITORCONFIG_EDITORCONFIG_CHECKER  # in Super-Linter
-  # - GHERKIN_GHERKIN_LINT  # in Super-Linter
-  - KUBERNETES_KUBECONFORM
-  - KUBERNETES_KUBEVAL  # in Super-Linter
-  - OPENAPI_SPECTRAL  # in Super-Linter
-  - PUPPET_PUPPET_LINT
-  # - SNAKEMAKE_LINT  # in Super-Linter
-  # - SNAKEMAKE_SNAKEFMT  # in Super-Linter
-  # - TEKTON_TEKTON_LINT  # in Super-Linter
-  - TERRAFORM_CHECKOV
-  - TERRAFORM_KICS
-  - TERRAFORM_TERRAFORM_FMT  # in Super-Linter
-  - TERRAFORM_TERRAGRUNT  # in Super-Linter
-  - TERRAFORM_TERRASCAN  # in Super-Linter
-  - TERRAFORM_TFLINT  # in Super-Linter
-
-  ### Code quality checkers
-  - COPYPASTE_JSCPD  # in Super-Linter
-  - REPOSITORY_CHECKOV
-  - REPOSITORY_DEVSKIM
-  - REPOSITORY_DUSTILOCK
-  - REPOSITORY_GITLEAKS  # in Super-Linter
-  - REPOSITORY_GIT_DIFF
-  - REPOSITORY_GOODCHECK
-  - REPOSITORY_SECRETLINT
-  - REPOSITORY_SEMGREP
-  - REPOSITORY_SYFT
-  - REPOSITORY_TRIVY
-  # - SPELL_CSPELL
-  - SPELL_MISSPELL
-  - SPELL_PROSELINT
-
-SHOW_ELAPSED_TIME: false
-FILEIO_REPORTER: false
-FAIL_IF_MISSING_LINTER_IN_FLAVOR: true
-# DISABLE_ERRORS: true # Uncomment if you want MegaLinter to detect errors but not block CI to pass
