--- lib/way_of_working/templates/.rubocop
@@ -1,1 +1,0 @@
---config .github/linters/rubocop_defaults.yml
