--- lib/way_of_working/templates/docs/decisions/README.md
@@ -1,7 +1,0 @@
-# Decisions
-
-This directory contains decision records for the project.
-
-For new ADRs, please use [adr-template.md](adr-template.md) as basis.
-More information on MADR is available at <https://adr.github.io/madr/>.
-General information about architectural decision records is available at <https://adr.github.io/>.
