+++ test/generators/rake_tasks/init_test.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
 require 'test_helper'
-require 'way_of_working/paths'
-require 'way_of_working/generators/rake_tasks/init'
 
 module WayOfWorking
   module Generators
