+++ test/git/summary_tag_test.rb
@@ -4,7 +4,6 @@ require 'active_support'
 require 'active_support/core_ext'
 
 require 'test_helper'
-require 'way_of_working/git/repo_reader'
 
 module WayOfWorking
   module Git
