+++ way_of_working.gemspec
@@ -42,4 +42,5 @@ Gem::Specification.new do |spec|
   spec.add_dependency 'nokogiri'
   spec.add_dependency 'rainbow', '~> 3.1'
   spec.add_dependency 'thor', '~> 1.2'
+  spec.add_dependency 'zeitwerk', '~> 2.6.18'
 end
